Simplyfing pr2_arm_kinematics_plugin, and deleting a bunch of no longer necessary files
