+++ constraint_samplers/CMakeLists.txt
@@ -15,5 +15,5 @@ install(TARGETS ${MOVEIT_LIB_NAME}
 install(DIRECTORY include/
   DESTINATION include)
 
-catkin_add_gtest(test_constraint_samplers test/test_constraint_samplers.cpp test/pr2_arm_kinematics_plugin.cpp test/pr2_arm_ik.cpp test/pr2_arm_ik_solver.cpp test/pr2_arm_kinematics_utils.cpp)
+catkin_add_gtest(test_constraint_samplers test/test_constraint_samplers.cpp test/pr2_arm_kinematics_plugin.cpp test/pr2_arm_ik.cpp)
 target_link_libraries(test_constraint_samplers ${MOVEIT_LIB_NAME} ${catkin_LIBRARIES})
