+++ constraint_samplers/test/pr2_arm_ik.h
@@ -37,13 +37,79 @@
 #include <Eigen/Core>
 #include <Eigen/LU>// provides LU decomposition
 #include <kdl/chainiksolver.hpp>
-
-#include "pr2_arm_kinematics_utils.h"
-#include "pr2_arm_kinematics_constants.h"
-
+#include <moveit_msgs/GetPositionFK.h>
+#include <moveit_msgs/GetPositionIK.h>
+#include <moveit_msgs/GetKinematicSolverInfo.h>
 
 namespace pr2_arm_kinematics
 {
+static const int NUM_JOINTS_ARM7DOF = 7;
+
+static const double IK_EPS = 1e-5;
+
+inline double distance(const urdf::Pose &transform)
+{
+  return sqrt(transform.position.x*transform.position.x+transform.position.y*transform.position.y+transform.position.z*transform.position.z);
+}
+
+
+inline bool solveQuadratic(const double &a, const double &b, const double &c, double *x1, double *x2)
+{
+  double discriminant = b*b-4*a*c;
+  if(fabs(a) < IK_EPS)
+  {
+    *x1 = -c/b;
+    *x2 = *x1;
+    return true;
+  }
+  //ROS_DEBUG("Discriminant: %f\n",discriminant);
+  if (discriminant >= 0)
+  {      
+    *x1 = (-b + sqrt(discriminant))/(2*a); 
+    *x2 = (-b - sqrt(discriminant))/(2*a);
+    return true;
+  } 
+  else if(fabs(discriminant) < IK_EPS)
+  {
+    *x1 = -b/(2*a);
+    *x2 = -b/(2*a);
+    return true;
+  }
+  else
+  {
+    *x1 = -b/(2*a);
+    *x2 = -b/(2*a);
+    return false;
+  }
+}
+
+inline bool solveCosineEqn(const double &a, const double &b, const double &c, double &soln1, double &soln2)
+{
+  double theta1 = atan2(b,a);
+  double denom  = sqrt(a*a+b*b);
+
+  if(fabs(denom) < IK_EPS) // should never happen, wouldn't make sense but make sure it is checked nonetheless
+  {
+#ifdef DEBUG
+    std::cout << "denom: " << denom << std::endl;
+#endif
+    return false;
+  }
+  double rhs_ratio = c/denom;
+  if(rhs_ratio < -1 || rhs_ratio > 1)
+  {
+#ifdef DEBUG
+    std::cout << "rhs_ratio: " << rhs_ratio << std::endl;
+#endif
+    return false;
+  }
+  double acos_term = acos(rhs_ratio);
+  soln1 = theta1 + acos_term;
+  soln2 = theta1 - acos_term;
+
+  return true;
+}
+
 class PR2ArmIK
 {
 public:
