+++ constraint_samplers/test/pr2_arm_kinematics_plugin.cpp
@@ -45,6 +45,203 @@ using namespace std;
 
 namespace pr2_arm_kinematics {
 
+bool PR2ArmIKSolver::getCount(int &count, 
+                              const int &max_count, 
+                              const int &min_count)
+{
+  if(count > 0)
+  {
+    if(-count >= min_count)
+    {   
+      count = -count;
+      return true;
+    }
+    else if(count+1 <= max_count)
+    {
+      count = count+1;
+      return true;
+    }
+    else
+    {
+      return false;
+    }
+  }
+  else
+  {
+    if(1-count <= max_count)
+    {
+      count = 1-count;
+      return true;
+    }
+    else if(count-1 >= min_count)
+    {
+      count = count -1;
+      return true;
+    }
+    else
+      return false;
+  }
+}
+
+PR2ArmIKSolver::PR2ArmIKSolver(const urdf::ModelInterface &robot_model, 
+                               const std::string &root_frame_name,
+                               const std::string &tip_frame_name,
+                               const double &search_discretization_angle, 
+                               const int &free_angle):ChainIkSolverPos()
+{
+  search_discretization_angle_ = search_discretization_angle;
+  free_angle_ = free_angle;
+  root_frame_name_ = root_frame_name;
+  if(!pr2_arm_ik_.init(robot_model,root_frame_name,tip_frame_name))
+    active_ = false;
+  else
+    active_ = true;
+}
+
+int PR2ArmIKSolver::CartToJnt(const KDL::JntArray& q_init, 
+                              const KDL::Frame& p_in, 
+                              KDL::JntArray &q_out)
+{
+  Eigen::Matrix4f b = KDLToEigenMatrix(p_in);
+  std::vector<std::vector<double> > solution_ik;
+  if(free_angle_ == 0)
+  {
+    ROS_DEBUG("Solving with %f",q_init(0)); 
+    pr2_arm_ik_.computeIKShoulderPan(b,q_init(0),solution_ik);
+  }
+  else
+  {
+    pr2_arm_ik_.computeIKShoulderRoll(b,q_init(2),solution_ik);
+  }
+  
+  if(solution_ik.empty())
+    return -1;
+
+  double min_distance = 1e6;
+  int min_index = -1;
+
+  for(int i=0; i< (int) solution_ik.size(); i++)
+  {     
+    ROS_DEBUG("Solution : %d",(int)solution_ik.size());
+
+    for(int j=0; j < (int)solution_ik[i].size(); j++)
+    {   
+      ROS_DEBUG("%d: %f",j,solution_ik[i][j]);
+    }
+    ROS_DEBUG(" ");
+    ROS_DEBUG(" ");
+
+    double tmp_distance = computeEuclideanDistance(solution_ik[i],q_init);
+    if(tmp_distance < min_distance)
+    {
+      min_distance = tmp_distance;
+      min_index = i;
+    }
+  }
+
+  if(min_index > -1)
+  {
+    q_out.resize((int)solution_ik[min_index].size());
+    for(int i=0; i < (int)solution_ik[min_index].size(); i++)
+    {   
+      q_out(i) = solution_ik[min_index][i];
+    }
+    return 1;
+  }
+  else
+    return -1;
+}
+
+int PR2ArmIKSolver::CartToJntSearch(const KDL::JntArray& q_in, 
+                                    const KDL::Frame& p_in, 
+                                    KDL::JntArray &q_out, 
+                                    const double &timeout)
+{
+  KDL::JntArray q_init = q_in;
+  double initial_guess = q_init(free_angle_);
+
+  ros::Time start_time = ros::Time::now();
+  double loop_time = 0;
+  int count = 0;
+
+  int num_positive_increments = (int)((pr2_arm_ik_.solver_info_.limits[free_angle_].max_position-initial_guess)/search_discretization_angle_);
+  int num_negative_increments = (int)((initial_guess-pr2_arm_ik_.solver_info_.limits[free_angle_].min_position)/search_discretization_angle_);
+  ROS_DEBUG("%f %f %f %d %d \n\n",initial_guess,pr2_arm_ik_.solver_info_.limits[free_angle_].max_position,pr2_arm_ik_.solver_info_.limits[free_angle_].min_position,num_positive_increments,num_negative_increments);
+  while(loop_time < timeout)
+  {
+    if(CartToJnt(q_init,p_in,q_out) > 0)
+      return 1;
+    if(!getCount(count,num_positive_increments,-num_negative_increments))
+      return -1;
+    q_init(free_angle_) = initial_guess + search_discretization_angle_ * count;
+    ROS_DEBUG("%d, %f",count,q_init(free_angle_));
+    loop_time = (ros::Time::now()-start_time).toSec();
+  }
+  if(loop_time >= timeout)
+  {
+    ROS_DEBUG("IK Timed out in %f seconds",timeout);
+    return TIMED_OUT;
+  }
+  else
+  {
+    ROS_DEBUG("No IK solution was found");
+    return NO_IK_SOLUTION;
+  }
+  return NO_IK_SOLUTION;
+}
+
+bool getKDLChain(const urdf::ModelInterface& model, const std::string &root_name, const std::string &tip_name, KDL::Chain &kdl_chain)
+{
+  // create robot chain from root to tip
+  KDL::Tree tree;
+  if (!kdl_parser::treeFromUrdfModel(model, tree))
+  {
+    ROS_ERROR("Could not initialize tree object");
+    return false;
+  }
+  if (!tree.getChain(root_name, tip_name, kdl_chain))
+  {
+    ROS_ERROR_STREAM("Could not initialize chain object for base " << root_name << " tip " << tip_name);
+    return false;
+  }
+  return true;
+}
+
+Eigen::Matrix4f KDLToEigenMatrix(const KDL::Frame &p)
+{
+  Eigen::Matrix4f b = Eigen::Matrix4f::Identity();
+  for(int i=0; i < 3; i++)
+  {
+    for(int j=0; j<3; j++)
+    {
+      b(i,j) = p.M(i,j);
+    }
+    b(i,3) = p.p(i);
+  }
+  return b;
+}
+
+double computeEuclideanDistance(const std::vector<double> &array_1, const KDL::JntArray &array_2)
+{
+  double distance = 0.0;
+  for(int i=0; i< (int) array_1.size(); i++)
+  {
+    distance += (array_1[i] - array_2(i))*(array_1[i] - array_2(i));
+  }
+  return sqrt(distance);
+}
+
+void getKDLChainInfo(const KDL::Chain &chain,
+                     moveit_msgs::KinematicSolverInfo &chain_info)
+{
+  int i=0; // segment number
+  while(i < (int)chain.getNrOfSegments())
+  {
+    chain_info.link_names.push_back(chain.getSegment(i).getName());
+    i++;
+  }
+}
+
 PR2ArmKinematicsPlugin::PR2ArmKinematicsPlugin():active_(false){}
 
 bool PR2ArmKinematicsPlugin::isActive()
@@ -86,7 +283,6 @@ bool PR2ArmKinematicsPlugin::initialize(const std::string& group_name,
   }
   else
   {
-
     pr2_arm_ik_solver_->getSolverInfo(ik_solver_info_);
     pr2_arm_kinematics::getKDLChainInfo(kdl_chain_,fk_solver_info_);
     fk_solver_info_.joint_names = ik_solver_info_.joint_names;
@@ -114,52 +310,7 @@ bool PR2ArmKinematicsPlugin::getPositionIK(const geometry_msgs::Pose &ik_pose,
                                            std::vector<double> &solution,
                                            moveit_msgs::MoveItErrorCodes &error_code) const
 {
-  if(!active_)
-  {
-    ROS_ERROR("kinematics not active");
-    error_code.val = error_code.NO_IK_SOLUTION; 
-    return false;
-  }
-    
-  KDL::Frame pose_desired;
-  Eigen::Affine3d tp;
-  tf::poseMsgToEigen(ik_pose, tp);
-  tf::transformEigenToKDL(tp, pose_desired);
-
-  //Do the IK
-  KDL::JntArray jnt_pos_in;
-  KDL::JntArray jnt_pos_out;
-  jnt_pos_in.resize(dimension_);
-  for(int i=0; i < dimension_; i++)
-  {
-    jnt_pos_in(i) = ik_seed_state[i];
-  }
-
-  int ik_valid = pr2_arm_ik_solver_->CartToJnt(jnt_pos_in,
-                                               pose_desired,
-                                               jnt_pos_out);
-  if(ik_valid == pr2_arm_kinematics::NO_IK_SOLUTION)
-  {
-    error_code.val = error_code.NO_IK_SOLUTION; 
-    return false;
-  }
-
-  if(ik_valid >= 0)
-  {
-    solution.resize(dimension_);
-    for(int i=0; i < dimension_; i++)
-    {
-      solution[i] = jnt_pos_out(i);
-    }
-    error_code.val = error_code.SUCCESS;
-    return true;
-  }
-  else
-  {
-    ROS_DEBUG("An IK solution could not be found");   
-    error_code.val = error_code.NO_IK_SOLUTION; 
-    return false;
-  }
+  return false;
 }
 
 bool PR2ArmKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose &ik_pose,
@@ -224,93 +375,7 @@ bool PR2ArmKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose &ik_pose
                                               std::vector<double> &solution,
                                               moveit_msgs::MoveItErrorCodes &error_code) const
 {
-  if(!active_)
-  {
-    ROS_ERROR("kinematics not active");
-    error_code.val = error_code.PLANNING_FAILED; 
-    return false;
-  }
-  KDL::Frame pose_desired;
-  Eigen::Affine3d tp;
-  tf::poseMsgToEigen(ik_pose, tp);
-  tf::transformEigenToKDL(tp, pose_desired);
-
-  //Do the IK
-  KDL::JntArray jnt_pos_in;
-  KDL::JntArray jnt_pos_out;
-  jnt_pos_in.resize(dimension_);
-  for(int i=0; i < dimension_; i++)
-  {
-    jnt_pos_in(i) = ik_seed_state[i];
-  }
-
-  unsigned int old_free_angle = pr2_arm_ik_solver_->getFreeAngle();
-  pr2_arm_ik_solver_->setFreeAngle(redundancy);
-  int ik_valid = pr2_arm_ik_solver_->CartToJntSearch(jnt_pos_in,
-                                                     pose_desired,
-                                                     consistency_limit,
-                                                     jnt_pos_out,
-                                                     timeout);
-  pr2_arm_ik_solver_->setFreeAngle(old_free_angle);
-
-  if(ik_valid == pr2_arm_kinematics::NO_IK_SOLUTION)
-  {
-    error_code.val = error_code.NO_IK_SOLUTION; 
-    return false;
-  }
-
-  if(ik_valid >= 0)
-  {
-    solution.resize(dimension_);
-    for(int i=0; i < dimension_; i++)
-    {
-      solution[i] = jnt_pos_out(i);
-    }
-    error_code.val = error_code.SUCCESS;
-    return true;
-  }
-  else
-  {
-    ROS_DEBUG("An IK solution could not be found");   
-    error_code.val = error_code.NO_IK_SOLUTION; 
-    return false;
-  }
-}
-
-
-void PR2ArmKinematicsPlugin::desiredPoseCallback(const KDL::JntArray& jnt_array, 
-                                                 const KDL::Frame& ik_pose,
-                                                 moveit_msgs::MoveItErrorCodes& error_code) const
-{
-  std::vector<double> ik_seed_state;
-  ik_seed_state.resize(dimension_);
-  for(int i=0; i < dimension_; i++)
-    ik_seed_state[i] = jnt_array(i);
-
-  geometry_msgs::Pose ik_pose_msg;
-  Eigen::Affine3d tp;
-  tf::transformKDLToEigen(ik_pose, tp);
-  tf::poseEigenToMsg(tp, ik_pose_msg);
-
-  desiredPoseCallback_(ik_pose_msg,ik_seed_state,error_code);
-}
-
-
-void PR2ArmKinematicsPlugin::jointSolutionCallback(const KDL::JntArray& jnt_array, 
-                                                   const KDL::Frame& ik_pose,
-                                                   moveit_msgs::MoveItErrorCodes& error_code) const
-{
-  std::vector<double> ik_seed_state;
-  ik_seed_state.resize(dimension_);
-  for(int i=0; i < dimension_; i++)
-    ik_seed_state[i] = jnt_array(i);
-
-  geometry_msgs::Pose ik_pose_msg;
-  Eigen::Affine3d tp;
-  tf::transformKDLToEigen(ik_pose, tp);
-  tf::poseEigenToMsg(tp, ik_pose_msg);
-
-  solutionCallback_(ik_pose_msg,ik_seed_state,error_code);
+  return false;
 }
 
 bool PR2ArmKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose &ik_pose,
@@ -320,50 +385,7 @@ bool PR2ArmKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose &ik_pose
                                               const IKCallbackFn &solution_callback,
                                               moveit_msgs::MoveItErrorCodes &error_code) const
 {
-  if(!active_)
-  {
-    ROS_ERROR("kinematics not active");
-    error_code.val = error_code.PLANNING_FAILED;
-    return false;
-  }
-  KDL::Frame pose_desired;
-  Eigen::Affine3d tp;
-  tf::poseMsgToEigen(ik_pose, tp);
-  tf::transformEigenToKDL(tp, pose_desired);
-
-  //  desiredPoseCallback_ = desired_pose_callback;
-  //  solutionCallback_    = solution_callback;
-
-  //Do the IK
-  KDL::JntArray jnt_pos_in;
-  KDL::JntArray jnt_pos_out;
-  jnt_pos_in.resize(dimension_);
-  for(int i=0; i < dimension_; i++)
-  {
-    jnt_pos_in(i) = ik_seed_state[i];
-  }
-
-  int ik_valid = pr2_arm_ik_solver_->CartToJntSearch(jnt_pos_in,
-                                                     pose_desired,
-                                                     jnt_pos_out,
-                                                     timeout,
-                                                     error_code,
-                                                     boost::bind(solution_callback,_1, _2, _3),
-                                                     boost::bind(solution_callback, _1, _2, _3));
-  if(ik_valid >= 0)
-  {
-    solution.resize(dimension_);
-    for(int i=0; i < dimension_; i++)
-    {
-      solution[i] = jnt_pos_out(i);
-    }
-    return true;
-  }
-  else
-  {
-    ROS_DEBUG("An IK solution could not be found");   
-    return false;
-  }
+  return false;
 }
 
 bool PR2ArmKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose &ik_pose,
@@ -375,101 +397,14 @@ bool PR2ArmKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose &ik_pose
                                               const IKCallbackFn &solution_callback,
                                               moveit_msgs::MoveItErrorCodes &error_code) const
 {
-  if(!active_)
-  {
-    ROS_ERROR("kinematics not active");
-    error_code.val = error_code.PLANNING_FAILED;
-    return false;
-  }
-  KDL::Frame pose_desired;
-
-  Eigen::Affine3d tp;
-  tf::poseMsgToEigen(ik_pose, tp);
-  tf::transformEigenToKDL(tp, pose_desired);
-
-  //  desiredPoseCallback_ = desired_pose_callback;
-  //  solutionCallback_    = solution_callback;
-
-  //Do the IK
-  KDL::JntArray jnt_pos_in;
-  KDL::JntArray jnt_pos_out;
-  jnt_pos_in.resize(dimension_);
-  for(int i=0; i < dimension_; i++)
-  {
-    jnt_pos_in(i) = ik_seed_state[i];
-  }
-
-  unsigned int old_free_angle = pr2_arm_ik_solver_->getFreeAngle();
-  pr2_arm_ik_solver_->setFreeAngle(redundancy);
-  int ik_valid = pr2_arm_ik_solver_->CartToJntSearch(jnt_pos_in,
-                                                     pose_desired,
-                                                     jnt_pos_out,
-                                                     timeout,
-                                                     consistency_limit,
-                                                     error_code,
-                                                     boost::bind(solution_callback, _1, _2, _3),
-                                                     boost::bind(solution_callback, _1, _2, _3));
-  pr2_arm_ik_solver_->setFreeAngle(old_free_angle);
-  if(ik_valid == pr2_arm_kinematics::NO_IK_SOLUTION)
-    return false;
-
-  if(ik_valid >= 0)
-  {
-    solution.resize(dimension_);
-    for(int i=0; i < dimension_; i++)
-    {
-      solution[i] = jnt_pos_out(i);
-    }
-    return true;
-  }
-  else
-  {
-    ROS_DEBUG("An IK solution could not be found");   
-    return false;
-  }
+  return false;
 }
 
 bool PR2ArmKinematicsPlugin::getPositionFK(const std::vector<std::string> &link_names,
                                            const std::vector<double> &joint_angles,
                                            std::vector<geometry_msgs::Pose> &poses) const
 {
-  if(!active_)
-  {
-    ROS_ERROR("kinematics not active");
-    return false;
-  }
-
-  KDL::Frame p_out;
-  KDL::JntArray jnt_pos_in;
-  geometry_msgs::PoseStamped pose;
-  tf::Stamped<tf::Pose> tf_pose;
-
-  jnt_pos_in.resize(dimension_);
-  for(int i=0; i < dimension_; i++)
-  {
-    jnt_pos_in(i) = joint_angles[i];
-    ROS_DEBUG("Joint angle: %d %f",i,joint_angles[i]);
-  }
-
-  poses.resize(link_names.size());
-
-  bool valid = true;
-  for(unsigned int i=0; i < poses.size(); i++)
-  {
-    ROS_DEBUG("End effector index: %d",pr2_arm_kinematics::getKDLSegmentIndex(kdl_chain_,link_names[i]));
-    if(jnt_to_pose_solver_->JntToCart(jnt_pos_in,p_out,pr2_arm_kinematics::getKDLSegmentIndex(kdl_chain_,link_names[i])) >=0)
-    {
-      Eigen::Affine3d tp;
-      tf::transformKDLToEigen(p_out, tp);
-      tf::poseEigenToMsg(tp, poses[i]);
-    }
-    else
-    {
-      ROS_ERROR("Could not compute FK for %s",link_names[i].c_str());
-      valid = false;
-    }
-  }
-  return valid;
+  return false;
 }
 
 const std::vector<std::string>& PR2ArmKinematicsPlugin::getJointNames() const
