+++ constraint_samplers/test/test_constraint_samplers.cpp
@@ -49,6 +49,7 @@
 #include <gtest/gtest.h>
 #include <urdf_parser/urdf_parser.h>
 #include <fstream>
+#include <boost/bind.hpp>
 
 #include "pr2_arm_kinematics_plugin.h"
